BOM implement

+++ external_pr_files/check_pr_and_update_branch.sh
@@ -35,16 +35,21 @@ if [ -z "$CHANGED_FILES" ]; then
     exit 0
 fi
 
+file_in_BOM {
+    local file=$1
+    grep -q "^$file$" "$BOM_FILE"
+}
 
 for file in $CHANGED_FILES; do
-    mkdir -p "$(dirname "$file")"
-    git show parent-repo/$PR_BRANCH:"$file" > "$file" 2>/dev/null
+    if file_in_BOM "$file"; then
+        mkdir -p "$(dirname "$file")"
+        git show parent-repo/$PR_BRANCH:"$file" > "$file" 2>/dev/null
 done
 
 PR_FILE_STATUS=$(gh pr view $PR_NUMBER --repo $GITHUB_REPOSITORY --json files --jq '.files[] | select(.status == "removed") | .path' 2>/dev/null || echo "")
 
 for deleted_file in $PR_FILE_STATUS; do
-    if [ -f "$deleted_file" ]; then
+    if file_in_BOM "$file" && [ -f "$deleted_file" ]; then
         rm "$deleted_file"
     fi
 done
